requirements.txt
streamlit
pandas
plotly
pdfplumber
openpyxl
statsmodels
fpdf2